oQo-scripts/requirements.txt
feedparser==6.0.11
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.13.4
python-dateutil==2.9.0
playwright==1.44.0
openai==1.84.0
psycopg2-binary==2.9.10